requirements.txt
Django>=5.0
pandas>=2.0.0
requests>=2.31.0
python-decouple>=3.8
gunicorn>=21.2.0
whitenoise>=6.6.0